Moved exchange back into separate module
